Migrate Demandable test to rspec
